add fotos speakers
